third_party/blink/web_tests/virtual/scoped-custom-element-registry/external/wpt/custom-elements/registries/Element-customElementRegistry-exceptions-expected.txt

This is a testharness.js-based test.
[FAIL] customElementRegistry on a failed custom element created by setting innerHTML should return the associated scoped registry
  assert_equals: expected object "[object CustomElementRegistry]" but got object "[object CustomElementRegistry]"
Harness: the test ran to completion.
